Refactor: view 클래스 입출력기준으로 분리
- view클래스의 메서드들을 입/출력을 기준으로해서 InputView와 OutputView로 분리함.
